feat: Milestone 2 — deploy scripts + frontend dashboard
Deploy scripts:
- script/Deploy.s.sol — mines CREATE2 hook salt, pre-computes all nonce-based
  addresses, deploys 6 contracts in correct dependency order with address verification
- script/DeployReactive.s.sol — deploys TridentReactive to Reactive Network
- script/InitPool.s.sol — initialises Uniswap v4 pool with hook attached
- .env.example — documents all required env vars with values and formulas

Frontend (Next.js + wagmi v2 + RainbowKit):
- Wagmi config with Unichain Sepolia chain definition
- Minimal ABIs for TridentHook, ILReserveVault, PositionTracker
- Live Fee Breakdown card — previewFee() with arb + boundary premium bars
- IL Reserve Vault card — health ratio, capture rate, reserve vs liability
- Reactive Network Status — primed deviation, gamma score, manipulation guard
- LP Position card — wallet-connected range lookup + projected vault payout
- Swap Activity feed — real-time SwapFeeBreakdown event stream

Bug fix: TridentReactive.sol type mismatch (int24 vs uint8 ternary)

+++ reactive/TridentReactive.sol
@@ -316,7 +316,7 @@ contract TridentReactive is AbstractPausableReactive {
         }
 
         nearestBoundary = nearest;
-        uint256 tickSpacingsAway = uint256(minDiff) / uint256(uint24(TICK_SPACING > 0 ? TICK_SPACING : 1));
+        uint256 tickSpacingsAway = uint256(minDiff) / uint256(uint24(TICK_SPACING > 0 ? TICK_SPACING : int24(1)));
         gammaScore = 1e18 / (tickSpacingsAway + 1);
     }
